c/topic_table: lift topic_metadata_item type
This commit lifts the topic_metadata_item type into its own "module".
A future commit will add a probe to the type and having it defined
in line is a bit cluttered.
