+++ src/v/cluster/CMakeLists.txt
@@ -104,6 +104,7 @@ v_cc_library(
     partition_balancer_planner.cc
     partition_balancer_backend.cc
     partition_balancer_rpc_handler.cc
+    topic_metadata_item.cc
   DEPS
     Seastar::seastar
     controller_rpc
