+++ src/v/cluster/topic_table.h
@@ -13,6 +13,7 @@
 
 #include "cluster/commands.h"
 #include "cluster/non_replicable_topics_frontend.h"
+#include "cluster/topic_metadata_item.h"
 #include "cluster/types.h"
 #include "model/fundamental.h"
 #include "model/limits.h"
@@ -41,53 +42,12 @@ public:
         cancel_requested,
         force_cancel_requested
     };
-    /**
-     * Replicas revision map is used to track revision of brokers in a replica
-     * set. When a node is added into replica set its gets the revision assigned
-     */
-    using replicas_revision_map
-      = absl::flat_hash_map<model::node_id, model::revision_id>;
 
     struct in_progress_update {
         std::vector<model::broker_shard> previous_replicas;
         in_progress_state state;
         model::revision_id update_revision;
         replicas_revision_map replicas_revisions;
-    };
-
-    struct topic_metadata_item {
-        topic_metadata metadata;
-        // replicas revisions for each partition
-        absl::node_hash_map<model::partition_id, replicas_revision_map>
-          replica_revisions;
-
-        bool is_topic_replicable() const {
-            return metadata.is_topic_replicable();
-        }
-
-        assignments_set& get_assignments() {
-            return metadata.get_assignments();
-        }
-
-        const assignments_set& get_assignments() const {
-            return metadata.get_assignments();
-        }
-        model::revision_id get_revision() const {
-            return metadata.get_revision();
-        }
-        std::optional<model::initial_revision_id> get_remote_revision() const {
-            return metadata.get_remote_revision();
-        }
-        const model::topic& get_source_topic() const {
-            return metadata.get_source_topic();
-        }
-
-        const topic_configuration& get_configuration() const {
-            return metadata.get_configuration();
-        }
-        topic_configuration& get_configuration() {
-            return metadata.get_configuration();
-        }
     };
 
     using delta = topic_table_delta;
